Move external models to API package
