+++ src/main/java/com/bnc/sbjb/model/api/ValidationError.java
@@ -1,4 +1,4 @@
-package com.bnc.sbjb.model;
+package com.bnc.sbjb.model.api;
 
 public class ValidationError extends SubError {
