+++ src/test/java/com/bnc/sbjb/model/api/TestSubError.java
@@ -1,4 +1,4 @@
-package com.bnc.sbjb.model;
+package com.bnc.sbjb.model.api;
 
 public class TestSubError extends SubError {
