+++ src/test/java/com/bnc/sbjb/model/api/ValidationErrorTest.java
@@ -1,7 +1,8 @@
-package com.bnc.sbjb.model;
+package com.bnc.sbjb.model.api;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.bnc.sbjb.model.api.ValidationError;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
