+++ src/test/java/com/bnc/sbjb/rest/DefaultErrorControllerTest.java
@@ -2,7 +2,7 @@ package com.bnc.sbjb.rest;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.bnc.sbjb.model.CustomError;
+import com.bnc.sbjb.model.api.CustomError;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
